♻️refactor: 커스텀 구분자 검증 및 추출 기능 모듈화
